requirements.txt
python-dotenv
streamlit
langchain
langchain-community
langchain-openai
langchain-groq
mysql-connector-python